#3: refactor: rename app swot to swot_item
