+++ backend/settings.py
@@ -37,7 +37,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'django.contrib.sites',
-    'swot.apps.SwotConfig',
+    'swot_item.apps.SwotConfig',
     'swot_item_vote.apps.SwotItemVoteConfig',
     'rest_framework',
     'authenticationjwt.apps.AuthenticationjwtConfig',
