+++ backend/urls.py
@@ -2,7 +2,7 @@ from django.conf.urls import url, include
 from django.contrib import admin
 from rest_framework.urlpatterns import format_suffix_patterns
 
-from swot import views as swotviews
+from swot_item import views as swotitemviews
 from swot_item_vote import views as voteviews
 
 URL_API_AUTH = r'^api/auth/'
@@ -23,12 +23,12 @@ urlpatterns = [
     ),
     url(
         URL_API_SWOT_ITEM,
-        swotviews.swot_list,
+        swotitemviews.swot_list,
         name='get_post_delete_swot_item',
     ),
     url(
         URL_API_SWOT_ITEM_DETAIL,
-        swotviews.swot_detail,
+        swotitemviews.swot_detail,
         name='delete_swot_item',
     ),
     url(
