+++ swot_item/apps.py
@@ -5,4 +5,4 @@ from django.apps import AppConfig
 
 
 class SwotConfig(AppConfig):
-    name = 'swot'
+    name = 'swot_item'
