+++ swot_item/fixtures/swotItems.json
@@ -1,6 +1,6 @@
 [
   {
-    "model": "swot.SwotItem",
+    "model": "swot_item.SwotItem",
     "pk": "1",
     "fields": {
       "created": "2017-09-04",
@@ -9,7 +9,7 @@
     }
   },
   {
-    "model": "swot.SwotItem",
+    "model": "swot_item.SwotItem",
     "pk": "2",
     "fields": {
       "created": "2017-09-04",
@@ -18,7 +18,7 @@
     }
   },
   {
-    "model": "swot.SwotItem",
+    "model": "swot_item.SwotItem",
     "pk": "3",
     "fields": {
       "created": "2017-09-04",
