+++ swot_item/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from swot.models import SwotItem
+from swot_item.models import SwotItem
 
 
 class SwotItemSerializer(serializers.ModelSerializer):
