+++ swot_item/tests/test_models.py
@@ -2,14 +2,14 @@ import datetime
 
 from django.test import TestCase
 
-from swot.models import SwotItem
+from swot_item.models import SwotItem
 from swot_item_vote.models import Vote
 
 ITEM_1_TEXT = 'Strength item #1'
 
 
 class ItemTestCase(TestCase):
-    fixtures = ['items.json']
+    fixtures = ['swotItems.json']
 
     def setUp(self):
         pass
